+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Added a `--skipTypecheck` CLI option to skip typechecking before tests.
+
 ## 16.1.2 - 2026-03-17
 
 ### Fixed